scafold complete

+++ Products/ProductHat.java
@@ -2,7 +2,7 @@ package Products;
 
 public class ProductHat extends Product{
 
-  public ProductHat(long id, double price, String title) {
+  public ProductHat(int id, double price, String title) {
     super(ProductCategory.Hat, id, price, title);
   }
 }
