+++ Products/ProductPant.java
@@ -2,7 +2,7 @@ package Products;
 
 public class ProductPant extends Product{
 
-  public ProductPant(long id, double price, String title) {
+  public ProductPant(int id, double price, String title) {
     super(ProductCategory.Pant, id, price, title);
   }
 }
