+++ Products/ProductShirt.java
@@ -2,7 +2,7 @@ package Products;
 
 public class ProductShirt extends Product{
 
-  public ProductShirt(long id, double price, String title) {
+  public ProductShirt(int id, double price, String title) {
     super(ProductCategory.Shirt, id, price, title);
   }
 }
