+++ Products/TypeConstructor.java
@@ -1,5 +1,5 @@
 package Products;
 
 public interface TypeConstructor {
-    Product apply(long id, double price, String title);
+    Product apply(int id, double price, String title);
 }
